dropout/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
numpy
scikit-learn
groq
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
